Moved HapiHL7ExpressionEvaluator and HapiHL7Message to rs-e2e project
