+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/ruleengine/AssertionRuleEngine.java
@@ -1,7 +1,7 @@
 package gov.hhs.cdc.trustedintermediary.rse2e.ruleengine;
 
 import ca.uhn.hl7v2.model.Message;
-import gov.hhs.cdc.trustedintermediary.external.hapi.HapiHL7Message;
+import gov.hhs.cdc.trustedintermediary.rse2e.external.hapi.HapiHL7Message;
 import gov.hhs.cdc.trustedintermediary.ruleengine.RuleLoader;
 import gov.hhs.cdc.trustedintermediary.ruleengine.RuleLoaderException;
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
